extract hook to custom hook for the sake of clarity
