+++ packages/example/src/components/modals/SubtitlesModal.tsx
@@ -1,11 +1,9 @@
-import { useEffect } from 'react';
 import { SpatialNavigationRoot } from '../../../../lib/src/spatial-navigation/components/Root';
 import { DefaultFocus } from '../../../../lib/src/spatial-navigation/context/DefaultFocusContext';
-import { useLockSpatialNavigation } from '../../../../lib/src/spatial-navigation/context/LockSpatialNavigationContext';
 import { Button } from '../../design-system/components/Button';
 import { Spacer } from '../../design-system/components/Spacer';
 import { GenericModal } from './GenericModal';
-import { EventArg, useNavigation } from '@react-navigation/native';
+import { useLockModal } from '../../hooks/useLockModal';
 
 interface SubtitlesModalProps {
   isModalVisible: boolean;
@@ -18,24 +16,7 @@ export const SubtitlesModal = ({
   setIsModalVisible,
   setSubtitles,
 }: SubtitlesModalProps) => {
-  const { lock, unlock } = useLockSpatialNavigation();
-  const navigation = useNavigation();
-
-  // Locks the parent navigator when the modal is open and unlocks it when it's closed
-  useEffect(() => {
-    if (isModalVisible) {
-      const navigationListener = (e: EventArg<'beforeRemove', true>) => {
-        e.preventDefault();
-        setIsModalVisible(false);
-      };
-      navigation.addListener('beforeRemove', navigationListener);
-      lock();
-      return () => {
-        navigation.removeListener('beforeRemove', navigationListener);
-        unlock();
-      };
-    }
-  }, [isModalVisible, lock, unlock, navigation, setIsModalVisible]);
+  useLockModal({ isModalVisible, setIsModalVisible });
 
   return (
     <SpatialNavigationRoot>
